Decouple packages - remove duplication
