+++ packages/@podman-desktop-companion/rpc/package.json
@@ -1,21 +1,16 @@
 {
-  "name": "@podman-desktop-companion/container-client",
+  "name": "@podman-desktop-companion/rpc",
   "version": "1.0.0",
   "author": "Ionut Stoica <stoica.ionut@gmail.com>",
-  "description": "Container desktop helper client",
+  "description": "Container desktop helper rpc",
   "main": "src/index.js",
+  "typings": "./src/index.d.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "license": "ISC",
   "dependencies": {
-    "@podman-desktop-companion/executor": "file:../executor",
-    "@podman-desktop-companion/terminal": "file:../terminal",
-    "@podman-desktop-companion/utils": "file:../utils",
-    "axios": "^0.24.0",
-    "electron-cfg": "^1.2.7",
     "electron-log": "^4.4.1",
-    "threads": "^1.7.0",
     "uuid": "^8.3.2"
   },
   "engines": {
