+++ packages/@podman-desktop-companion/terminal/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@podman-desktop-companion/terminal",
+  "version": "1.0.0",
+  "author": "Ionut Stoica <stoica.ionut@gmail.com>",
+  "description": "Container desktop helper terminal",
+  "main": "src/index.js",
+  "typings": "./src/index.d.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "license": "ISC",
+  "dependencies": {
+    "@podman-desktop-companion/executor": "file:../executor"
+  },
+  "engines": {
+    "node": "~16"
+  },
+  "repository": "https://github.com/iongion/podman-desktop-companion.git"
+}
