+++ packages/@podman-desktop-companion/utils/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@podman-desktop-companion/utils",
+  "version": "1.0.0",
+  "author": "Ionut Stoica <stoica.ionut@gmail.com>",
+  "description": "Container desktop helper utils",
+  "main": "src/index.js",
+  "typings": "./src/index.d.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "license": "ISC",
+  "dependencies": {},
+  "peerDependencies": {
+    "axios": "^0.24.0"
+  },
+  "engines": {
+    "node": "~16"
+  },
+  "repository": "https://github.com/iongion/podman-desktop-companion.git"
+}
