+++ packages/@podman-desktop-companion/utils/src/index.d.ts
@@ -1,0 +1,4 @@
+declare module "@podman-desktop-companion/utils" {
+  import { AxiosRequestConfig } from "axios";
+  export declare function axiosConfigToCURL(opts: AxiosRequestConfig): string;
+}
