+++ shell/package.json
@@ -8,6 +8,8 @@
   "main": "./public/main.js",
   "dependencies": {
     "@podman-desktop-companion/container-client": "file:../packages/@podman-desktop-companion/container-client",
+    "@podman-desktop-companion/rpc": "file:../packages/@podman-desktop-companion/rpc",
+    "@podman-desktop-companion/terminal": "file:../packages/@podman-desktop-companion/terminal",
     "ajv": "8.8.2",
     "axios": "^0.24.0",
     "electron-cfg": "^1.2.7",
